WP4_Thor-main/service/search_engine/server/src/requirements.txt
gunicorn[gevent]
Flask
Werkzeug
python-dotenv
flask-cors
elasticsearch
flasgger
pyyaml
openai
flask_jwt_extended
pymongo
